+++ tests/data/VersionedItem.php
@@ -1,0 +1,47 @@
+<?php
+
+namespace yii2tech\tests\unit\ar\softdelete\data;
+
+use yii\db\ActiveRecord;
+use yii2tech\ar\softdelete\SoftDeleteBehavior;
+
+/**
+ * @property int $id
+ * @property string $name
+ * @property bool $isDeleted
+ * @property int $deletedAt
+ * @property int $version
+ */
+class VersionedItem extends ActiveRecord
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function behaviors()
+    {
+        return [
+            'softDelete' => [
+                'class' => SoftDeleteBehavior::className(),
+                'softDeleteAttributeValues' => [
+                    'isDeleted' => true
+                ],
+            ],
+        ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public static function tableName()
+    {
+        return 'Item';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function optimisticLock()
+    {
+        return 'version';
+    }
+}